power: add partial-resume framework
Partial resume refers to the concept of not waking up userspace when the
kernel
comes out of suspend for certain types of events that we wish to
discard.  An
example is a network packet that can be disacarded in the kernel, or
spurious
wakeup event that we wish to ignore.  Partial resume allows drivers to
register
callbacks, one one hand, and provides hooks into the PM's suspend/resume
mechanism, on the other.

When a device resumes from suspend, the core suspend/resume code invokes
partialresume to check to see if the set of wakeup interrupts all have
matching handlers. If this is not the case, the PM subsystem can
continue to
resume as before.  If all of the wakeup sources have matching handlers,
then
those are invoked in turn (and can block), and if all of them reach
consensus
that the reason for the wakeup can be ignored, they say so to the PM
subsystem,
which goes right back into suspend.  This latter support is implemented
in a
separate change.

Signed-off-by: Iliyan Malchev <malchev@google.com>
Change-Id: Id50940bb22a550b413412264508d259f7121d442